+++ packages/scim/src/scim.test.ts
@@ -250,7 +250,7 @@ describe("SCIM", () => {
 				        {
 				          "description": "A Boolean value indicating the User's administrative status.",
 				          "multiValued": false,
-				          "mutability": "readOnly",
+				          "mutability": "readWrite",
 				          "name": "active",
 				          "required": false,
 				          "returned": "default",
@@ -503,7 +503,7 @@ describe("SCIM", () => {
 				    {
 				      "description": "A Boolean value indicating the User's administrative status.",
 				      "multiValued": false,
-				      "mutability": "readOnly",
+				      "mutability": "readWrite",
 				      "name": "active",
 				      "required": false,
 				      "returned": "default",